Update patch set 1
Patch Set 1:

— gFran

About 2, I understand it’s not easy to get the env in Unity and until, they find an easier way, I’m ok with having the explicit init method if it is impossible to to get the env out in Unity. Or is It  just confusing but it can be done? In the latter case I would get rid of it.

I definitely prefer to have an init method that doesn’t require any parameter other than the env for two reasons:

1) It’s consistent with how other google APIs work (eg Firebase)

2) It makes no sense for an external user that parameters that are internal to Android need to be passed to an Android library: it’s bad and confusing design.

Patch-set: 1
Tag: mailMessageId=<1D714BB4-699A-470D-89E1-7723AA586158@google.com>