SwappyVk: implement SwappyVkAndroidFallback
Implement SwappyVkAndroidFallback path which will be used
if VK_GOOGLE_display_timing is not available.

Change Cube sample main loop to be able to work with SwappyVk.
SwappyVk uses Choreographer, and Choreographer runs its callback in
app's Looper context, which means that app needs to wait on
ALooper_pollAll() to be able to wakeup from a Choreographer callback.

Test: manual
Change-Id: Id09c3bb65b139124dec54f3878d1673cec9da176