Update patch set 10
Patch Set 10: Code-Review-1

Sorry, but I don't understand the motivation for this change.  Is this a change that others will want (e.g. do we want to push this upstream)?  Have you tested that it builds and works fine on Windows and Linux?

Patch-set: 10
Reviewer: Gerrit User 1098333 <1098333@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=-1