Update patch set 10
Patch Set 10:

> Patch Set 10: Code-Review-1
> 
> Sorry, but I don't understand the motivation for this change.  Is this a change that others will want (e.g. do we want to push this upstream)?  Have you tested that it builds and works fine on Windows and Linux?

This is fixing a strange behavior on Android Studio. Without it, shader code will not get compiled unless Build->Refresh Linked C++ Projects is done.

Please note that this CMakeLists.txt is not the same as the one on VulkanTools, but a different versions composed to work on Android.

Patch-set: 10
Reviewer: Gerrit User 1388190 <1388190@85c56323-6fa9-3386-8a01-6480fb634889>