Update patch set 8
Patch Set 8:

Hi Ady,

LGTM for what you're trying to do.  In practice, how much does it add to each frame?

Could you also do an approach where it's not completely random?  Let me describe what I'm after.  Let's say we have two parameters: 1) a steady-state amount of per-frame time to add; and 2) a variable amount per-frame time to add.

There are times I'd like to experiment with different steady-state times on different devices (e.g. a 60Hz, 90Hz, and 120Hz display, like a Razer), to look for boundary conditions (that might push us over a vsync boundary).  Once we know the boundary, it would be good to then start varying each frame so that some frames go past the boundary and others don't go past the boundary.  There are applications that have this problem on some devices, which cause them to get jank because some of their frames are on-time and others late (or some are on time and some early).  I want to see how each algorithm works in those cases.  Both algorithms may have problems that we're not anticipating now (i.e. that's why we test things:-).

What you have here seems like a good approach for the variable part of what I want.  Perhaps repeating the logic for the non-variable part (and not giving it a random number) is all we need.

Thanks,
Ian

Patch-set: 8
Reviewer: Gerrit User 1098333 <1098333@85c56323-6fa9-3386-8a01-6480fb634889>