Update patch set 4
Patch Set 4:

> Patch Set 4:
> 
> > Patch Set 4:
> > 
> > (1 comment)
> > 
> > The code LGTM but I'm confused by the overall direction here. I thought we were going to have samples/common/src rather than src/common? If it's src/common then I'm guessing we intend to compile this into the SDK itself. I think that's fine but if so we should rename common to gamesdk_common or something because some of the game engines might have /common on their own include paths. WDYT?
> 
> Trace.h is used internally in the Swappy and TuningFork libraries, not just samples. We don't have gamesdk_include, so why do we need gamesdk_common? The swappy and tuningfork code shouldn't be compiled with any external include directories, even if they are part of a game engine. Or maybe I'm missing something here?

What I'm suggesting (and I think Fran and Ady are also saying) is that we should have 'internal' versions of Trace/Log which are used by Game SDK and then a second 'common' version which is essentially a support library for building the samples. This is ~sensible because it forces the samples to mimic what we can expect when integrating into external engines. For this to be workable I think that the files should have different names and function prefixes in order to disambiguate them.

This approach will not work for the Settings.h case which I commented on below (as in that case I think we legitimately want the samples to call into something that the game sdk does not yet expose).

Patch-set: 4
Reviewer: Gerrit User 1269249 <1269249@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=0