Update patch set 2
Patch Set 2: Code-Review+2

(3 comments)

Gave +2 because this is a big change and the majority of it looks good. I'm assuming we're blocked from using device_info until the next update from g3? 

I think that the code in tuningfork_extra.cpp and tuningfork_internal.h probably needs to be moved around a bit so it's a bit easier to navigate, but I'm happy for that to be in a follow up CL.

Patch-set: 2
Reviewer: Gerrit User 1269249 <1269249@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+2