Update patch set 1
Patch Set 1: Code-Review+2

This is definitely an improvement, so let's put it in. But let's also make sure we cover the header/cpp relationships and layout in the upcoming API review. It feels like it could still be cleaner, but a code diff is not the best way to think about it.

Patch-set: 1
Reviewer: Gerrit User 1269249 <1269249@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+2