Update patch set 2
Patch Set 2:

(1 comment)

> Patch Set 2:
> 
> > Patch Set 2:
> > 
> > > Patch Set 2:
> > > 
> > > > Patch Set 2:
> > > > 
> > > > I think we should abandon building shared objects and then the other Swappy versioning CL is not needed. It's addressing the problem of mismatches between headers and so's, which would go away. Is that correct, Ady, Dan? We can then add a similar version function for Swappy, as Florian's added here for Tuning Fork, so we can track usage.
> > > 
> > > Even if we use a static library we might have ABI issues if there is a mistmatch between the header that application is using and the one the library uses. 
> > > 
> > > This CL adds a version to the .so, but how does it provide runtime feedback in case of a mismatch ?  Do we expect the application to call to TUNINGFORK_VERSION_SYMBOL() ?
> > 
> > This CL does not indeed provide any feedback at all in case of mismatch - my initial goal was just to be able to track in apks the usage of TuningFork, and its version ;)
> > 
> > If we want to also have runtime feedback in case of mismatch, I'll add a runtime check similar to the one done in CL947000 (i.e: make the init function of TuningFork a static inline function in the header, that is calling an internal init function passing the version number, which is then checked against the version numbers that were put at the time of the compilation of the library).
> > 
> > Do you think it's ok for me to proceed and add this to this CL? Or in another CL?
> 
> I see. We can use this CL to add all the version related stuff, no need for a different CL.

Just to chime in on the topic of "what happens if you have a .h/.a mismatch"? The behaviour will depend on the type of mismatch, but in general you get a linker error. For some changes you can get undefined behaviour though (i.e. change the size of a struct but not the name) - in practice that maps to memory truncation or stomps. 

However this is the same as you get with any statically linked tech. WWise, Havok etc. all use this approach already. This kind of developer is familiar with such risks. 

We do not expect the application to call TUNINGFORK_VERSION_SYMBOL() directly. It is called as a side-effect of TuningFork_init(). I think we should employ exactly the same approach with Swappy?

Patch-set: 2