Add CMake function to uncomment to develop Game SDK from samples
This reuse the idea from aosp/955228, but with a single line change to do in the CMakeList.txt of each sample - and documented in the README.
When uncommenting add_gamesdk_source, it is possible to have sources of Swappy/Tuning Fork integrated to Android Studio with debugger support.

There were issues with the "external" path, it seemed to be defined in a hacky way and was failing when compiling from samples.
Using now the same way to resolve path as for "third-party" folder.

Test: Running bouncyball from Android studio with and without add_gamesdk_source
Test: Tested adding a breakpoint in Swappy sources from bouncyball

Test: Running tftestapp from Android studio with and without add_gamesdk_source
Test: Tested adding a breakpoint in Tuning Fork sources from tftestapp

Test: checked that command line compilation still works:
Test: `cd samples/tuningfork/tftestapp && ./gradlew clean assemble && cd -`
Test: `cd samples/bouncyball && ./gradlew clean assemble && cd -`
Test: `cd samples/cube && ./gradlew clean assemble && cd -`
Test: `./gradlew cleanPath archiveTfZip unityZip`

Bug: 136197599

Change-Id: I156a32745465e0a385ab17457f5a7af65fd06ce2