Update patch set 2
Patch Set 2: Code-Review+2

> Patch Set 2:
> 
> Would like your mind about this kind of unit tests in game sdk (compiling/running outside of a Android environment).

It's great that you've added this type of test. I was talking to Ady about adding this sort of test for some of the auto-mode logic in Swappy and there's scope for unit tests of some of the Tuning Fork code too.

Patch-set: 2
Reviewer: Gerrit User 1373878 <1373878@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+2