Fix crash in GameSdkDeviceInfoJni when the native library is missing
HIGH_LEVEL_CL_EXPLANATION:
- The native library game_sdk_device_info_jni is not included in all flavors
of phonesky (e.g: x86). In this case, the Java library calling it should have been
robust against this and just logging the error about the library being missing.

- The error handling was not catching properly java.lang.UnsatisfiedLinkError
(which is an Error and not an Exception), resulting in a crash at the loading
time (see b/140598211).

- This is fixed by catching all potential Error (only java.lang.UnsatisfiedLinkError
is expected to be thrown according to the docs, but for safety we're catching
everything). We could catch Throwable instead of catching separately Exception and Error but having the two separated is a way to explicitly show that we do want to swallow these.

CHANGES_AFFECTING_LIVE_EXPERIMENTS:
- GameSdkDeviceInfo (enable_collection) - but this experiment is in dogfood and disabled for now

NON_LIVE_EXPERIMENT_GUARDED_CHANGES:
- None

UNGUARDED_CHANGES_AND_REASON:
- None

SHOULD_BE_NEUTRAL_WITH_RELEASE_BECAUSE:
- The experiment is in dogfood and disabled. This is adding additional error handling which will prevent a crash on devices having phonesky without the native library.

Test: Reproduced the crash on a Pixel 2 XL by running Phonesky with "game_sdk_device_info_jni" renamed to simulate the missing library
Test: Verified that with the patch, we're now just getting a line of log: "DeviceConfigurationHelperImpl.createGameSdkData: GameSdkDeviceInfo native library was not loaded, exception or error is: java.lang.UnsatisfiedLinkError: dalvik.system.PathClassLoader[...]" and no crash of the Play Store app.
PiperOrigin-RevId: 268716346
Change-Id: I70bb76378ad17e9c2b6853acaf2c7f8a6ce1584a