Update patch set 1
Patch Set 1:

(8 comments)

> Patch Set 1:
> 
> (8 comments)
> 
> Posted some comments. The callbacks needs to be renamed in the code as well, right ?

I don't think callbacks need to be renamed because, even if I added typedefs to make the declaration of the type of callbacks in SwappyTracer clearer, the names of the members themselves are unchanged.

The rest of the code, that is setting them to either nullptr or a C function (or a static method in C++), is only referring to the member names. For example:

  SwappyTracer tracer;
  tracer.preWait = swappy_trace_test_preWait;

with swappy_trace_test_preWait being a C function here. But the types are unchanged by the typedefs being added.

Adding fcarucci@ to reviewers to get clarification on currentFrameTimeStampMillis documentation.

Patch-set: 1