Adding CMake to PATH for Bender build
Adding CMake to PATH to resolve issue of CMake not found.
Alternate (probably simpler) solution to aosp/1151235,
previously avoided in the off-chance that modifying
PATH in the build.sh would have side effects. However,
aosp/1151235 has the issue of modifying the
local.properties forcefully and making it difficult to
build/run directly on Android Studio (instead of just
through gradlew). Thus, this is being tried.

Test: Forrest run P10211662, local testing with gradlew
and on Android Studio still building correctly.

Change-Id: I99587a33851087163382633f491d27ffcd827763