Add screen_refresh_rate functions
Add functions to set screen refresh rate, modified from aosp/1122259
For now we call to set refresh rate in the main thread VulkanDrawFrame(),
so we only need to call setScreenRefreshRateMainThread(), which is exactly
setScreenRefreshRateImpl().

In future development we would be having seperate thread(s) that does rendering.
In that case, we would need looper to attach the set refresh rate call to
the main thread, which refers to what setScreenRefreshRate() does.

Given currently I have only Pixel3/3a that runs inly at 60Hz so I could not
really tell if I set it successfully. Will get a Pixel4 which runs 90Hz at
default and see if this works.

Test: Build and run on Pixel3a with no crash
Change-Id: I1cc84d065ad4cc3d7f9b5ae7b5e36de71056d5d4