Add GAMESDK_ROOT_DIR argument to add_gamesdk_target
This allows to use this function to build a sample without compiling the
gamesdk sources, by using some prebuilt version of gamesdk being in
GAMESDK_ROOT_DIR

Add error message if the paths seems obviously wrong.

Updated comments and removed an unecessary definition of GAMESDK_LIB_DIR
in the parent scope.

Test: ran samples
Test: copied bouncyball out of the gamesdk (along with common), set
ROOT_DIR to point to the existing gamesdk with compiled binaries,
checked that the sample is compiling.

Change-Id: Ia73cfdb66ca35d862905475fb712b3b7ef0cf0f4