Add an extra sanity check for the existence of gamesdk library
When using add_gamesdk_target without building the gamesdk from sources,
it's easy to use a SDK/NDK version that is not matching the ones we're
providing. Add an explicit check for this to fail fast.

Test: ./gradlew clean assemble for tftestapp, bouncyball, cube.
Test: Copied bouncyball and common from samples directory in a new,
unrelated directory. Changed add_gamesdk_target in bouncyball
CMakeLists.txt to specify the ROOT_DIR. Verify that compilation works
and app can run.

Bug: 143345931

Change-Id: Ia2655efa4bbfee51539fe6e52b1e62c2e4ed4b06