Add samples to packaged game sdk
Copy the samples (only for Swappy for now) when packaging the game sdk,
and use a CMakeLists that use the packaged gamesdk, instead of building
from sources.

Samples can be opened and compiled without changes if you have the
proper SDK/NDK. If not, you'll get a "friendly" error message telling
that your SDK/NDK is not suppored.

Test: create a local build with "./gradlew archiveZip". Extract it somewhere. Open with Android Studio bouncyball or cube that are inside the samples folder of the extracted zip. Compile (after verifying that the proper SDK/NDK versions are used).
Change-Id: Iba1c739ac84c608c622e190667bde4a228226d41