Update patch set 1
Patch Set 1: Code-Review+1

Thanks for the changes :) 

In the case of DO_LOCAL_BUILD, should we default to CMAKE_BUILD_TYPE the GAMESDK_BUILD_TYPE to avoid building a Release version when you're building stuff for Debug?
Or is this not a problem?

Patch-set: 1
Reviewer: Gerrit User 1498014 <1498014@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+1