Revised crash when report file is open.
It was programmatic and assuming that such possibility should never
occur. That was wrong: if MainActivity is destroyed by Android OS and
re-created afterwards, the new MainActivity will try to re-open the
already open report and that's not an abnormal situation.
Just a matter of asking whether is already open before trying.

Test: I tested this manually, both keeping back stack activities and not
keeping these.

Change-Id: I63c3048ac2d2f5fa47bf11668b4bebaf59ce2832