Adding color to Material
Adding diffuse color attribute to Material.
Color is removed from the vertex format (and relevant values in
Polyhedron are removed), and color will be provided to the fragment
shader through the material UBO rather than from the vertex data.

Material may have optional input Texture or diffuse color,
with default conditions for both. Default texture is a 1x1 white
square, and Texture is refactored to provide a constructor
that will create this texture without having an asset file for it.

Test: Build/run on Android Studio + Pixel 3a

Change-Id: I2a499d5a5e8836834c37b6f3010787eb3db53494