Update patch set 2
Patch Set 2: Code-Review+1

(8 comments)

Hey!
A few comments on this:

Can we tie these into the systrace markers? That is, calling start/end section should call the Systrace function as well, so you only need to do one (or the other) and not both.

It'd be great if you could add a scope-based way of calling start/stopSection (see comment in the header file). That way we can also avoid people typing the section name wrong when it closes.

Might be nice to add categories of events. E.g. an Animation category, or a Post-Processing category. Categories could have a color, a name, and an ID - and you'd pass the category into start/stop. It'd act as a bucket for events. Categories should also have an optional associated budget - so you can render a warning onscreen if the game goes over budget in any given frame.

Si

Patch-set: 2
Reviewer: Gerrit User 1566081 <1566081@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+1