Update patch set 4
Patch Set 4:

The general problem I have with this design is that every time a measure is started, it creates an Event object in the heap, which can very much lead to a lot of fragmentation in the heap.
I would try to redesign in a way that events are not created on the heap but just allocated from an array of events that can eventually just grow.

Patch-set: 4
Reviewer: Gerrit User 1376737 <1376737@85c56323-6fa9-3386-8a01-6480fb634889>