Remove cube sample check for NDK dir
Remove the check for the NDK env variable in cube/app build.gradle
because this seems non standard, does not play well with Android Studio
(you have to manually define it) and is not actually needed (this looks
more like a sanity check, but not adapted to Android Studio).

Test: Build with "ANDROID_HOME=../prebuilts/sdk ./gradlew gamesdkZip",
extract the zip somewhere, open bouncyball and cube samples, run them
from Android Studio.
Test: checked that tftestapp, cube and bouncyball works when building
them from sources.

Change-Id: Ib2e61104f4096b88b8d610c272b9ad5a7b11dc47