Add android.ndkVersion to samples build.gradle
This specify the NDK version to use for samples, as it's a best practice
to define it.

Use 20.0.5594570 as it's the most up-to-date version in the 20.0
releases (and this is the version that we have as a prebuilt NDK).

Updated Gradle plugin to 3.5 (so that ndkVersion is not ignored). Had to update to Gradle 5.4 at the same time and updated root build.gradle for compatibility with Gradle 5.

Test: Build with "ANDROID_HOME=../prebuilts/sdk ./gradlew gamesdkZip",
extract the zip somewhere, open bouncyball and cube samples, run them
from Android Studio.
Test: checked that tftestapp, cube and bouncyball works when building
them from sources.

Change-Id: I2c9fc227af76f54329c1e1d6d34fd5cd92600f2f