Add android.ndkVersion to samples build.gradle
This specify the NDK version to use for samples, as it's a best practice
to define it. If missing on the developer computer, Android Studio
should offer to install it.

Use 19.2.5345600 as it's the most up-to-date version in the 19.x
version.

Also remove the check for the NDK env variable in cube/app build.gradle
because this seems non standard, does not play well with Android Studio
(you have to manually define it) and not actually needed (this looks
more like a sanity check). As we now have android.ndkVersion, this seems
redundant.

Test: Build with "ANDROID_HOME=../prebuilts/sdk ./gradlew gamesdkZip",
extract the zip somewhere, open bouncyball and cube samples, run them
from Android Studio.
Test: checked that tftestapp, cube and bouncyball works when building
them from sources.

Bug: 144694251

Change-Id: I2c9fc227af76f54329c1e1d6d34fd5cd92600f2f