Set vetex memory offset for drawString()
In the previous commit the successing drawString() calls would overwrite
data of previous calls, which is caused by not having set an offset.
This commit adds this basic but crucial part to enable multiple drawString
calls in one render pass. In the start of each frame we will clear the offset
to start reading strings that are to be rendered and write the relating
data to the vertex memory. The size of the vertex buffer is now 120 character.

Test: Build and test on Pixel 3a
Change-Id: I0d2d0e88de999077c1fa72549f49ad6db6094044