Remove cube.cpp source file
Cube has a C++ version in cube.cpp, but Swappy is not integrated and this file is not used for compiling the sample (the C version only is used).
This creates confusion, so removing this file. If we want to add a demo with Cube in C++ and Swappy, we can add this back later.

Test: compile cube
Change-Id: I63367fb6e204a9aefa3ff8648760ccbe8d65abc8