Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Can this violate licence?
> 
> As a side note: Cube causes too much trouble with its licence, and its not a good sample for developers to follow,  shall we create another Vulcan sample Tetrahedron on our own?

I contacted opensource-licensing asking this exact question (you're CC'ed so you should get their answer too).
In the case we can't remove it, we can still add a notice at the top of this file that it's unused (but would still be source of confusion) - to be confirmed with opensource-licensing too.

Patch-set: 1
Reviewer: Gerrit User 1498014 <1498014@85c56323-6fa9-3386-8a01-6480fb634889>