Update patch set 9
Patch Set 9: Presubmit-Ready+1

Updated Gradle plugin to 3.5 (so that ndkVersion is not ignored). Had to update to Gradle 5.4 at the same time.
Confirmed that locally, Android Studio is using this NDK and erroring if it's not installed.

Giving a try at the pre-submits.

Patch-set: 9
Reviewer: Gerrit User 1498014 <1498014@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=+1