Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> That's the idea right now yea, just loop through them all. You can make it faster by creating a spatial bounding hierarchy where you organize the buttons into buckets by dividing the screen into subsections so you only test buttons in the subsection that was tapped
> 
> It's not too complicated, but that could be a v2 feature (and it might actually be faster to just iterate through all of them if we have a small number of buttons haha)
> 
> That's the only way I know of atm to make it faster though, I'm always open to ideas!

Hmm, I guess that's fair, we could keep it that way for now. At any rate, I think having a Font object associated with the button might be helpful for a text button? Unless you're thinking of this in terms of like the part-of-screen tapping that's happening for +1/-1 polyhedron right now (making the Font optional?)

Patch-set: 1