Update patch set 1
Patch Set 1:

Also, thinking more about it, because our only option is to loop through buttons and check, plus no feedback is being returned from the button to the caller, any degree of overlap of two buttons could result in both buttons being pressed. This is not intrinsically a problem, but I think that having a 1:1 between user input and action is probably good to maintain? Which would require either letting the caller choose which button is pressed (based on proximity) or provide some return that says "this button actually ran the action, don't run anything else"

Patch-set: 1