Ensure samples are building Game SDK sources with the proper NDK
When samples are built from sources, they are launching a local build of
Game SDK. As we're now able to specify the NDK version in Gradle
(ndkVersion), we must pass the version to use to the Game SDK
build.gradle and use this NDK version (we have to search in side-by-side
NDK for the one with the specified version).

Test: cd samples/tuningfork/tftestapp && ./gradlew clean assemble
Test: tested with:
Test: * ndkVersion set (which will be used),
Test: * ndkVersion not set (so the default NDK is used),
Test: * ndkVersion not set but ANDROID_NDK_VERSION passed to add_gamesdk_target
Test: * ANDROID_NDK_VERSION passed with an incorrect value: error about
NDK not being found.

Test: ./gradlew archiveTfZip
Test: ANDROID_HOME=`pwd`/../prebuilts/sdk ./gradlew gamesdkZip

Change-Id: I9055c59eb4db3f4c2e885beb06771b8e136c48d0