Update patch set 1
Patch Set 1: Code-Review+1

(2 comments)

This is looking great, Shamyl!

Will clearly impact my change on the works but we have no choice here: this must happen.

I made a comment on marching cubes suite handler (an antipattern that likely happens in other places as well). But is not an error that your change introduced so I marked it self-resolved. Yet, would be great if you address this as part of the overall improvement.

Patch-set: 1
Reviewer: Gerrit User 1590267 <1590267@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+1