Fix crash in GetDebuggable when ApplicationInfo is minified
When compiling in release mode, there was a crash due to calling
GetIntField on a Java object with a null clz. This is because
GetDebuggable is calling flags() on an ApplicationInfo object that is
invalid.

ApplicationInfo object was invalid because not found when class names
are minified. We should handle this case properly to avoid a crash.

Test: added signingConfigs for release build of tftestapp, compiled
tftestapp in release, launched it, verified that the app runs properly.

Change-Id: I00d9b96084d6df05e47e5444a221793633404429