Fixes an error where the report log was erroneously closed
My understanding/analysis of the problem was that MainActivity could,
in some case, be destroyed while a test was running. In doing so the
report log would be closed. On finish of the running test it would
be re-opened, and any attempts to write to the report log failed.

The solution I've implemented is to *never* close the log; just
flush it. Closing the log has no purpose; it should always be
open, but flushed opportunistically.

After this fix we've gone from ~50-60 crashes on FTL (yikes!)
to < 10, and none of these were related to writing to a closed log.

Test: Tested on FTL
Change-Id: I35e4c2096fddcc792bef1390cca4e8922afb2865