CPU Temperature collection.
Implementation of a mechanism that detects system files that track CPU
temperatures. Until Android Q, there was no official way to estimate
such temperature. Still with Android Q, it can be said that not all
devices implement the novel Thermal API.
For that reason, we scan system folders to look for potential files that
hold such information.

As an extra, I refactored a bit System.cpp as it grew way to
heterogeneous, breaking the single-responsibility principle in a gross
manner.

Test: I tested this running it on FTL.

Change-Id: I385f66fa6e2bbcf54b8740518982d6a744c91ed5