Wait for Pi calculation
Heats up the device to measure performance under load. Can be
configured to use several different throttling mechanisms.

Test: The operation runs for the specified amount of time and
generates a report.json file. The number of threads listed in
that file should match the number of threads on the device,
and each thread should report how many iterations of the Pi
approximation it completed per interval.

Change-Id: Ib688c3c8ecca2f12712a89861c2249cd49f5120a