Update patch set 7
Patch Set 7:

(1 comment)

Is your plan to add a new test mode after mmapAnonTestActive, that uses mmapping the way I would expect games to use it, which is to mmap a file in read only mode?

I would not expect this to show much functional difference to using malloc, which displays the same behavior with respect to memory not being effectively consumed until written to.

Patch-set: 7
Reviewer: Gerrit User 1003913 <1003913@85c56323-6fa9-3386-8a01-6480fb634889>