Smooth out camera motion
Smooth out camera motion for moving forward/backward and strafing.
Currently, using button on_hold is unreliable b/c the button
signal is irregular and infrequent.

This does bring up whether we should change on_hold entirely;
we would be able to avoid this problem if we had
a loop where every frame we ran any on_hold functions that
were applicable. The only downside of that approach is
if there are uses of on_hold that do not require per-frame
execution (for example if we used on_hold to set the new bools
in this CL that would result in a lot of redundant variable
setting).

Test: build/run on Android Studio.
Change-Id: Ib1e78c5072e2861289db48fa9800e03978397021