Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> How materials are laid out and loaded is our choice. It shouldn't be in bender-kit.

For mesh_helpers it's more of a code representation of the MTL format found in the obj spec (and the ParseMTL is used to read the mtl file itself)

So that code isn't doing any sort of transformation of the materials into a bender-specific representation - I figured that the way that mesh_helpers treats MTL files was similar in style to the obj_loader treated OBJ files which is also in bender-kit

Or are we saying that the fact that it IS a representation of the OBJ spec is an implicit decision of how we want our data represented for bender and as a result it should be moved out of bender-kit? In that case we should also move the obj_loader code as well?

Patch-set: 2