Update patch set 3
Patch Set 3:

> Patch Set 3: Code-Review-2
> 
> This removes the NDK r14 and r15 builds of the Games SDK, not just the tests.

NDK r14 and r15 doesn't work with CMake 3.10. And they are broken after we upgrade cmake.

Can you confirm that r14 and r15 builds are still important? If that's the case we may have to pin the cmake to 3.6 for gamesdk.

Patch-set: 3
Reviewer: Gerrit User 1307957 <1307957@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=0