Modified Bender for GDC Demo
My current state of Bender for the GDC Demo.

Modifies camera position to be middle of the room, disables
movement buttons (only rotation w/ touch screen), and
disables front-to-back sorting for performance reasons.

Includes a new thread for loading OBJs for both init and resume.
I've added a loading screen also. There are also mutexes for
modifying the loading text and modifying/accessing meshes in
the render_graph.

Change-Id: Iff4cb2409edc4e90c88ee7deb224a0f94b64997a