Fix startup + some device loss crashes
Fix various crashes.

Not sure if this one is still relevant, but startup crashes, as a result
of some of the timing library not playing well with multithreading;
I've taken out the Timing in the threads for now and we can add them back when
we've figured out how to support the threading.

Also, fixed benderkit::Device code by adding DestroySurface;
since we've managed to avoid destroying the device when
we lose focus, we also don't destroy the swapchain and surface,
which used to be in ~Device. This leads to incorrect ref count
in the ANativeWindow.

Refactored OBJ loading code into LoadDemoModels().

Added code to pause loading (after completing the current OBJ) on
focus lost, and resume when we get the window back.

Change-Id: I5b7ca4f5172b95f14d7eaf969acaaf80332b995a