Fix startup + some device loss crashes
Fix startup crashes, which was a result of some of the timing library
not playing well with multithreading; I've taken out the Timing in
the threads for now and we can add them back when we've figured out
how to support the threading.

Fix some device loss crash, with Material StaticCleanup (b/c the
loading takes long and the default material may not have been
created yet, it should not be accessed unless it exists).

There's still crashing which I'm investigating, but primarily seems
to be things unable to Cleanup() themselves on device loss.

Change-Id: I5b7ca4f5172b95f14d7eaf969acaaf80332b995a