Fix startup + some device loss crashes
Fix startup crashes, which was a result of some of the timing library
not playing well with multithreading; I've taken out the Timing in
the threads for now and we can add them back when we've figured out
how to support the threading.

Fix some device loss crash, with Material StaticCleanup (b/c the
loading takes long and the default material may not have been
created yet, it should not be accessed unless it exists).

Also fixed some crashing on loading interrupt due to partial loading
of OBJs; I've opted to clean out the geometries/meshes if loading
is only partial, unless it has completed fully at least once,
since then all the resources necessary to OnResume the entire scene
will be available. There's still some crashing on loading interrupt,
however, which is hard to diagnose.

Change-Id: I5b7ca4f5172b95f14d7eaf969acaaf80332b995a