ASTC Compressed Textures
This code enables the use of ASTC textures.

A script has been written (convert_astc.sh) to iterate
through our current set of assets and convert all
.png's to .astc's; however, it is dependent on two
sets of libraries (ARM's astc-encoder
to do the actual encoding and ImageMagick to flip
the png's vertically prior to encoding, to counteract
the inversion by astcenc). These will be added to Chef
in addition to mesh loading and decimation after we
move to the team repo.

Changes include adding ASTC loading to texture.cc/h
and modifying utils/src/obj_loader.cc/h to take an
override mtl file (which is what we have currently been
using, just not as a param). CAUTION: this CL will
change the mtl name to _astc.mtl, which is an MTL
I created which names astc textures for its objects.
I may gate this depending on how we think it would be
appropriate to do so.

Test: I have an APK at x20/teams/android-games-devtech
named bender-v1.5-astc-12x12.apk, which uses
12x12 compressed textures. You may notice some materials
still appear odd, such as some of the Windows obj's,
or the fact that the floor is significantly brighter.
I believe this to be how we do Material and enforce
that all of the attributes are present, and will work
to figure out how to make these look normal.

Change-Id: Ie7442d0b118d8245593c87b90160bba0aeb91589