Launcher can be configured in tests.json.
The idea is that anything that make sense to be changed on a per-test
dimension level can be done in tests.json rather than in a constant.

This allows useful things like testing variations of builds, running
tests with different Firebase configurations like timeout, and comparing
performance in portrait and landscape mode.

Test: Manual.
Change-Id: I939936efd97ef4de5795917a53af5eb62b74c94b