Improved score reports.
The table header has multiple rows and uses rowspan to reduce
repetition in the header names. It is *far* easier to grok complex
test runs in this form. It also allows space for a more
straightforward labeling scheme to be used (converting the params to
a label).

The grouping of the best test scores has been generalized. Every
unique parameter combination (bar 'heuristic' - which is what is
being compared) forms its own group.

Test: Manual.
Change-Id: I50c46140f1bdafefa4349bf4047d76d69f2a9188