Heuristics driven by the JSON params and can have params themselves.
See bug for justification - I need configuration to be very flexible at
this stage of development, and need to be able to compare and tune the
input parameters.

Also adding my latest work on the heuristics including heuristics that
are based on the data from the discover tool.

I moved tryAlloc definition to MainActivity to accompany the other
native methods.

Bug: 154102290
Test: Manual
Change-Id: Id399b4839f65949d0ee13ae80d1d619695806c9b