Buffer-less font rendering
This is the initial version of font rendering without using a vertex buffer
and index buffer. By utilizing uniform buffer objects and push constants,
we can render about a hundred characters per draw call depeneding on the
size of the push constants available on the device. The implementation can
be seperated into these parts:

- Set up vertex empty input state in the pipeline
- Set up push constant range in the pipeline
- Change culling and front face settings for rasterization
- Push font data to a UBO
- Add data parsing and assembling logic to the vertex shader

What is further coming:
- Add empty vertex format to shader_state and vertex_format
- Support splitting long strings into multiple draw calls

Test: Build and run on Pixel 4 XL via Android Studio

Change-Id: I8fdf2d2c34353afd32e3f0627a1a53410ae48308